WIP: implement engine rpc
- Implement basic RPC functionality
- Implement JWT authentication

To run, change the variable ErigonDatadir to the correct value for your
machine.
Assumes that Erigon is currently running on localhost with he default
ports

Notes: Currently, sending UpdateForkChoice hangs indefinitely.
My best guess is that this is because I'm sending a dummy blockhash,
and the execution client is waiting to receive that block via p2p.
Needs investigation.