Implement loop to drive Erigon
Features:
- Create and authenticate an Engine RPC connection with Erigon
- Use the engine API to request blocks from the execution client, then add them to chain.  This is enough to run transactions via metamask
- Unit test

To run, change the variable ErigonDatadir to the correct value for your
machine.
Due to the lack of an underlying DB, this implementation assumes that Erigon starting from Genesis.
It also assumes that Erigon is currently running on localhost with the default ports and the `sovereign` chain.
To start erigon with this config: `erigon --chain=sovereign --datadir=~/sovereign --http`.